# packages/server/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ICategory": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"ownerId": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"ownerId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCategoryDto": {
				"properties": {
					"ownerId": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"ownerId",
					"name"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/category": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICategory"
								}
							}
						}
					}
				},
				"description": "Create a new category",
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Data to create category",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCategoryDto",
								"description": "Data to create category"
							}
						}
					}
				}
			}
		},
		"/category/{id}": {
			"delete": {
				"operationId": "DeleteById",
				"responses": {
					"200": {
						"description": "Успешность удаления",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"description": "Удалить категорию по ID",
				"security": [],
				"parameters": [
					{
						"description": "Идентификатор категории",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:8080/api/v1"
		}
	]
}